requirements.txt
opencv-python>=4.8,<5
numpy>=1.24
deepface>=0.0.79
requests>=2.31,<3
PyQt6>=6.6,<7
pygame>=2.5,<3
pyserial>=3.5,<4
spotipy>=2.23,<3
flask>=2.3,<3
python-dotenv>=1.0,<2
scikit-learn>=1.3,<2
